// src/interfaces/deafualt_rf_log.interface.ts
export interface DefaultRfLog {
  tg_id: string;
  rf_type: string;
  date_logged: string;
  date_solved: string;
  rf_status: number;
  staff_id: string;
  contact_method: string;
  contact_person: string;
  staff_id_solved: string;
  imei: string;
  app_version: string;
  rf_id: string;
  hub_id: string;
}